Refactor signers
