+++ index.js
@@ -10,7 +10,7 @@ const {
 } = require('./request/builder');
 
 const {IllegalArgumentError} = require('./illegal_argument_error');
-const {SignatureSigner} = require('./signer');
+const {EllipticSigner, WebCryptoSigner} = require('./signer');
 
 /**
  * This class handles all client interactions including registering certificates
@@ -60,7 +60,11 @@ class ClientServiceBase {
     }
 
     /** @const */
-    this.signer = new SignatureSigner(this.privateKeyPem);
+    if (typeof process === 'object') { // Node.js environment
+      this.signer = new EllipticSigner(this.privateKeyPem);
+    } else {
+      this.signer = new WebCryptoSigner(this.privateKeyPem);
+    }
 
     /**
      * The LedgerClient generated by gRPC library
