+++ signer/index.js
@@ -1,0 +1,9 @@
+const {EllipticSigner} = require('./EllipticSigner');
+const {JsrsasignSigner} = require('./JsrsasignSigner');
+const {WebCryptoSigner} = require('./WebCryptoSigner');
+
+module.exports = {
+  EllipticSigner,
+  JsrsasignSigner,
+  WebCryptoSigner,
+};
