+++ test/signature_signer.test.js
@@ -1,4 +1,4 @@
-const {SignatureSigner, JsrsasignSignatureSigner} = require('../signer.js');
+const {EllipticSigner, JsrsasignSigner} = require('../signer');
 const expect = require('chai').expect;
 const assert = require('chai').assert;
 const jsrsasign = require('jsrsasign');
@@ -8,12 +8,12 @@ const fs = require('fs');
 const chai = require('chai');
 chai.use(require('chai-as-promised'));
 
-describe('Class SignatureSigner', () => {
+describe('The Signature Signers', () => {
   afterEach(function() {
     sinon.restore();
   });
 
-  describe('Class JsrsasignSignatureSigner', () => {
+  describe('Class JsrsasignSigner', () => {
     // Mock the ArrayBuffertoHex in signer class
     beforeEach(function() {
       sinon.replace(jsrsasign, 'ArrayBuffertohex',
@@ -21,28 +21,27 @@ describe('Class SignatureSigner', () => {
     });
 
     it('throws exception when pem file is not correct', async () => {
-      const signer = new JsrsasignSignatureSigner('incorrect pem');
+      const signer = new JsrsasignSigner('incorrect pem');
       await expect(signer.sign('content'))
           .to.be.rejectedWith('Failed to load private key');
     });
 
     it('throws exception when it is not an ASN.1 hex string', async () => {
-      const signer = new JsrsasignSignatureSigner('mocked non ASN.1 string');
+      const signer = new JsrsasignSigner('mocked non ASN.1 string');
       await expect(signer.sign('content'))
           .to.be.rejectedWith('Failed to load private key');
     });
 
     it('should work properly', async () => {
       const pem = fs.readFileSync(__dirname + '/key.pem', 'utf-8');
-      const signer = new JsrsasignSignatureSigner(pem);
+      const signer = new JsrsasignSigner(pem);
       assert.instanceOf(await signer.sign('Content'), Uint8Array);
     });
   });
 
-  describe('Class EllipticSignatureSigner', () => {
+  describe('Class EllipticSigner', () => {
     it('throws exception when the pem file is not correct', async () => {
-      // EllipticSignature is used by default
-      const signer = new SignatureSigner('incorrect pem');
+      const signer = new EllipticSigner('incorrect pem');
 
       await expect(signer.sign('content')).to.be.rejectedWith(
           'Failed to sign the request'
@@ -50,7 +49,7 @@ describe('Class SignatureSigner', () => {
     });
     it('should work properly', async () => {
       const pem = fs.readFileSync(__dirname + '/key.pem', 'utf-8');
-      const signer = new SignatureSigner(pem);
+      const signer = new EllipticSigner(pem);
       assert.instanceOf(await signer.sign('Content'), Uint8Array);
     });
   });
